fix(frontend): vite8.x 文件循环依赖导致引用报错 #17365 # Reviewed, transaction id: 79252
