+++ dbm-ui/frontend/.env.production
@@ -1,4 +1,6 @@
 # .env.production 生产模式生效
 VITE_AJAX_URL_PREFIX=/
+
 VITE_ROUTER_PERFIX=/
+
 VITE_PUBLIC_PATH=/static/
