+++ dbm-ui/frontend/package.json
@@ -68,7 +68,7 @@
     "url-join": "5.0.0",
     "vue": "3.5.33",
     "vue-content-loader": "2.0.1",
-    "vue-i18n": "11.3.2",
+    "vue-i18n": "11.4.0",
     "vue-request": "2.0.4",
     "vue-router": "5.0.6",
     "vue-types": "6.0.0",
@@ -90,7 +90,7 @@
     "@vue/eslint-config-prettier": "10.2.0",
     "@vue/eslint-config-typescript": "14.7.0",
     "@vue/tsconfig": "0.9.1",
-    "commitlint": "20.5.0",
+    "commitlint": "20.5.2",
     "cross-env": "10.1.0",
     "cypress": "15.14.1",
     "cypress-multi-reporters": "2.0.5",
