+++ dbm-ui/frontend/src/layout/components/database-manage/components/module-group/Oracle.vue
@@ -91,7 +91,7 @@
 
   import { ClusterTypes, DBTypes } from '@common/const';
 
-  import { toolboxMenuList } from '@views/db-manage/oracle/toolbox/Index.vue';
+  import { toolboxMenuList } from '@views/db-manage/oracle/toolbox/toolboxMenuList';
 
   import CountTag from './components/CountTag.vue';
   import MenuGroup from './components/MenuGroup.vue';
