+++ dbm-ui/frontend/src/views/db-manage/oracle/toolbox/toolboxMenuList.ts
@@ -1,0 +1,32 @@
+/*
+ * TencentBlueKing is pleased to support the open source community by making 蓝鲸智云-DB管理系统(BlueKing-BK-DBM) available.
+ *
+ * Copyright (C) 2017-2023 THL A29 Limited, a Tencent company. All rights reserved.
+ *
+ * Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://opensource.org/licenses/MIT
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
+ * the specific language governing permissions and limitations under the License.
+ */
+
+import { TicketTypes } from '@common/const';
+
+import { t } from '@locales/index';
+
+export const toolboxMenuList = [
+  {
+    children: [
+      {
+        dbConsoleValue: 'oracle.toolbox.sqlExecute',
+        id: TicketTypes.ORACLE_EXEC_SCRIPT_APPLY,
+        name: t('变更SQL执行'),
+        parentId: 'sql',
+      },
+    ],
+    icon: 'db-icon-mysql',
+    id: 'sql',
+    name: t('SQL任务'),
+  },
+];
