+++ dbm-ui/frontend/src/views/db-manage/tendb-cluster/toolbox/toolboxMenuList.ts
@@ -1,0 +1,245 @@
+/*
+ * TencentBlueKing is pleased to support the open source community by making 蓝鲸智云-DB管理系统(BlueKing-BK-DBM) available.
+ *
+ * Copyright (C) 2017-2023 THL A29 Limited, a Tencent company. All rights reserved.
+ *
+ * Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://opensource.org/licenses/MIT
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
+ * the specific language governing permissions and limitations under the License.
+ */
+
+import { TicketTypes } from '@common/const';
+
+import { t } from '@locales/index';
+
+export const toolboxMenuList = [
+  {
+    children: [
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.sqlExecute',
+        id: TicketTypes.TENDBCLUSTER_IMPORT_SQLFILE,
+        name: t('变更SQL执行'),
+        parentId: 'spider_sql',
+      },
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.dbRename',
+        id: TicketTypes.TENDBCLUSTER_RENAME_DATABASE,
+        name: t('DB 重命名'),
+        parentId: 'spider_sql',
+      },
+    ],
+    icon: 'db-icon-mysql',
+    id: 'spider_sql',
+    name: t('SQL任务'),
+  },
+  {
+    children: [
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.masterSlaveSwap',
+        id: TicketTypes.TENDBCLUSTER_MASTER_SLAVE_SWITCH,
+        name: t('主从互切'),
+        parentId: 'spider_cluster_maintain',
+      },
+      {
+        bind: [TicketTypes.TENDBCLUSTER_MASTER_FAIL_OVER, TicketTypes.TENDBCLUSTER_INSTANCE_FAIL_OVER],
+        dbConsoleValue: 'tendbCluster.toolbox.instanceFailover',
+        id: TicketTypes.TENDBCLUSTER_MASTER_FAIL_OVER,
+        name: t('主库故障切换'),
+        parentId: 'spider_cluster_maintain',
+      },
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.capacityChange',
+        id: TicketTypes.TENDBCLUSTER_NODE_REBALANCE,
+        name: t('集群容量变更'),
+        parentId: 'spider_cluster_maintain',
+      },
+      {
+        bind: [
+          TicketTypes.TENDBCLUSTER_SPIDER_CONF_UP_DOWN,
+          TicketTypes.TENDBCLUSTER_SPIDER_ADD_NODES,
+          TicketTypes.TENDBCLUSTER_SPIDER_REDUCE_NODES,
+          TicketTypes.TENDBCLUSTER_SPIDER_SWITCH_NODES,
+        ],
+        dbConsoleValue: 'mysql.toolbox.proxyScaleUp',
+        id: TicketTypes.TENDBCLUSTER_SPIDER_ADD_NODES,
+        name: t('接入层变更'),
+        parentId: 'migrate',
+      },
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.masterSlaveClone',
+        id: TicketTypes.TENDBCLUSTER_MIGRATE_CLUSTER,
+        name: t('迁移主从'),
+        parentId: 'spider_cluster_maintain',
+      },
+      {
+        bind: [TicketTypes.TENDBCLUSTER_RESTORE_LOCAL_SLAVE, TicketTypes.TENDBCLUSTER_RESTORE_SLAVE],
+        id: TicketTypes.TENDBCLUSTER_RESTORE_LOCAL_SLAVE,
+        name: t('重建从库'),
+        parentId: 'spider_cluster_maintain',
+      },
+      {
+        id: TicketTypes.TENDBCLUSTER_CLUSTER_STANDARDIZE,
+        name: t('集群标准化'),
+        parentId: 'spider_cluster_maintain',
+      },
+      {
+        bind: [
+          TicketTypes.TENDBCLUSTER_LOCAL_UPGRADE,
+          TicketTypes.TENDBCLUSTER_SPIDER_UPGRADE,
+          TicketTypes.TENDBCLUSTER_REMOTE_UPGRADE,
+        ],
+        id: TicketTypes.TENDBCLUSTER_LOCAL_UPGRADE,
+        name: t('版本升级'),
+        parentId: 'spider_cluster_maintain',
+      },
+    ],
+    icon: 'db-icon-cluster',
+    id: 'spider_cluster_maintain',
+    name: t('集群维护'),
+  },
+  {
+    children: [
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.proxySlaveApply',
+        id: TicketTypes.TENDBCLUSTER_SPIDER_SLAVE_APPLY,
+        name: t('部署只读接入层'),
+        parentId: 'spider_entry',
+      },
+    ],
+    icon: 'db-icon-entry',
+    id: 'spider_entry',
+    name: t('访问入口'),
+  },
+  {
+    children: [
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.addMnt',
+        id: TicketTypes.TENDBCLUSTER_SPIDER_MNT_APPLY,
+        name: t('添加运维节点'),
+        parentId: 'spider_mnt',
+      },
+    ],
+    icon: 'db-icon-jiankong',
+    id: 'spider_mnt',
+    name: t('运维节点管理'),
+  },
+  {
+    children: [
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.dbTableBackup',
+        id: TicketTypes.TENDBCLUSTER_DB_TABLE_BACKUP,
+        name: t('库表备份'),
+        parentId: 'spider_copy',
+      },
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.dbBackup',
+        id: TicketTypes.TENDBCLUSTER_FULL_BACKUP,
+        name: t('全库备份'),
+        parentId: 'spider_copy',
+      },
+    ],
+    icon: 'db-icon-copy',
+    id: 'spider_copy',
+    name: t('备份'),
+  },
+  {
+    children: [
+      // {
+      //   dbConsoleValue: 'tendbCluster.toolbox.rollbackRecord',
+      //   id: 'spiderRollbackRecord',
+      //   name: t('构造实例'),
+      //   parentId: 'spider_fileback',
+      // },
+      {
+        bind: [TicketTypes.TENDBCLUSTER_FIXPOINT_EXIST, TicketTypes.TENDBCLUSTER_FIXPOINT_NEW],
+        dbConsoleValue: 'tendbCluster.toolbox.fixpoint',
+        id: TicketTypes.TENDBCLUSTER_FIXPOINT_EXIST,
+        name: t('构造'),
+        parentId: 'fileback',
+      },
+      {
+        bind: [TicketTypes.TENDBCLUSTER_FLASHBACK, TicketTypes.TENDBCLUSTER_ROLLBACK],
+        dbConsoleValue: 'tendbCluster.toolbox.flashback',
+        id: TicketTypes.TENDBCLUSTER_ROLLBACK,
+        name: t('回档'),
+        parentId: 'spider_fileback',
+      },
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.rollback',
+        id: TicketTypes.TENDBCLUSTER_ROLLBACK_CLUSTER,
+        name: t('定点构造（旧）'),
+        parentId: 'spider_fileback',
+      },
+    ],
+    icon: 'db-icon-rollback',
+    id: 'spider_fileback',
+    name: t('构造和回档'),
+  },
+  {
+    children: [
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.dbClear',
+        id: TicketTypes.TENDBCLUSTER_TRUNCATE_DATABASE,
+        name: t('清档'),
+        parentId: 'spider_data',
+      },
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.checksum',
+        id: TicketTypes.TENDBCLUSTER_CHECKSUM,
+        name: t('数据校验修复'),
+        parentId: 'spider_data',
+      },
+    ],
+    icon: 'db-icon-data',
+    id: 'spider_data',
+    name: t('数据处理'),
+  },
+  {
+    children: [
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.clientPermissionClone',
+        id: TicketTypes.TENDBCLUSTER_CLIENT_CLONE_RULES,
+        name: t('客户端权限克隆'),
+        parentId: 'spider_privilege',
+      },
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.dbInstancePermissionClone',
+        id: TicketTypes.TENDBCLUSTER_INSTANCE_CLONE_RULES,
+        name: t('DB实例权限克隆'),
+        parentId: 'spider_privilege',
+      },
+    ],
+    icon: 'db-icon-clone',
+    id: 'spider_privilege',
+    name: t('权限克隆'),
+  },
+  {
+    children: [
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.openareaTemplate',
+        id: TicketTypes.TENDBCLUSTER_OPEN_AREA,
+        name: t('开区模版'),
+        parentId: 'spider_openarea',
+      },
+    ],
+    icon: 'db-icon-template',
+    id: 'spider_openarea',
+    name: t('克隆开区'),
+  },
+  {
+    children: [
+      {
+        dbConsoleValue: 'tendbCluster.toolbox.webconsole',
+        id: 'SpiderWebconsole',
+        name: 'Webconsole',
+        parentId: 'spider_data_query',
+      },
+    ],
+    icon: 'db-icon-search',
+    id: 'spider_data_query',
+    name: t('数据查询'),
+  },
+];
